requirements.txt
itsdangerous
fastapi
uvicorn
python-dotenv